+++ server/src/api/services/internal/competitions.ts
@@ -1,11 +1,11 @@
 import { keyBy, mapValues, omit, uniqBy } from 'lodash';
 import moment from 'moment';
 import { Op, Sequelize } from 'sequelize';
+import { Competition, Group, Participation, Player } from '@models';
 import * as deltaService from '@services/internal/deltas';
 import * as groupService from '@services/internal/groups';
 import * as playerService from '@services/internal/players';
 import * as snapshotService from '@services/internal/snapshots';
-import { Competition, Group, Participation, Player } from '../../../database';
 import { ALL_METRICS, COMPETITION_STATUSES } from '../../constants';
 import { BadRequestError, NotFoundError } from '../../errors';
 import { durationBetween, isPast, isValidDate } from '../../util/dates';
