+++ server/src/api/services/internal/efficiency.ts
@@ -1,7 +1,7 @@
+import { Snapshot } from '@models';
 import { BOSSES, SKILLS } from '../../constants';
-import { getValueKey } from '../../util/metrics';
-import Snapshot from '../../modules/snapshots/snapshot.model';
 import mainAlgorithm from '../../modules/efficiency/algorithms/main';
+import { getValueKey } from '../../util/metrics';
 
 function calculateEHP(snapshot: Snapshot): number {
   // TODO: always use main ehp, for now
