+++ server/src/api/services/internal/snapshots.ts
@@ -2,8 +2,8 @@ import csv from 'csvtojson';
 import { keyBy, mapValues } from 'lodash';
 import moment from 'moment';
 import { Op } from 'sequelize';
+import { Snapshot } from '@models';
 import * as efficiencyService from '@services/internal/efficiency';
-import { Snapshot } from '../../../database';
 import { ACTIVITIES, ALL_METRICS, BOSSES, PERIODS, SKILLS } from '../../constants';
 import { BadRequestError, ServerError } from '../../errors';
 import { getMeasure, getRankKey, getValueKey } from '../../util/metrics';
