+++ server/src/database/index.ts
@@ -1,16 +1,18 @@
 import { Sequelize } from 'sequelize-typescript';
-import Achievement from '../api/modules/achievements/achievement.model';
-import Competition from '../api/modules/competitions/competition.model';
-import Participation from '../api/modules/competitions/participation.model';
-import InitialValues from '../api/modules/deltas/initialValues.model';
-import Group from '../api/modules/groups/group.model';
-import Membership from '../api/modules/groups/membership.model';
-import Player from '../api/modules/players/player.model';
-import Record from '../api/modules/records/record.model';
-import Snapshot from '../api/modules/snapshots/snapshot.model';
+import {
+  Achievement,
+  Competition,
+  Group,
+  InitialValues,
+  Membership,
+  NameChange,
+  Participation,
+  Player,
+  Record,
+  Snapshot
+} from '@models';
 import { isTesting } from '../env';
 import config from './config';
-import NameChange from './models/NameChange';
 
 const models = [
   Achievement,
@@ -31,16 +33,4 @@ const sequelize = new Sequelize({
   dialect: isTesting() ? 'sqlite' : 'postgres'
 });
 
-export {
-  sequelize,
-  Achievement,
-  Competition,
-  Group,
-  InitialValues,
-  Membership,
-  Participation,
-  Player,
-  Record,
-  Snapshot,
-  NameChange
-};
+export { sequelize };
