+++ server/src/database/models/InitialValues.ts
@@ -1,11 +1,10 @@
-import { BelongsTo, Column, CreatedAt, DataType, ForeignKey, Table } from 'sequelize-typescript';
-import { Player } from '../../../database';
-import HiscoreValues from '../deltas/hiscoreValues.model';
+import { BelongsTo, Column, DataType, ForeignKey, Table, UpdatedAt } from 'sequelize-typescript';
+import { HiscoresValues, Player } from '@models';
 
 // Define other table options
 const options = {
-  modelName: 'snapshots',
-  updatedAt: false,
+  modelName: 'initialValues',
+  createdAt: false,
   indexes: [
     {
       unique: true,
@@ -13,15 +12,12 @@ const options = {
     },
     {
       fields: ['playerId']
-    },
-    {
-      fields: ['createdAt']
     }
   ]
 };
 
 @Table(options)
-export default class Snapshot extends HiscoreValues {
+export default class InitialValues extends HiscoresValues {
   @Column({ type: DataType.INTEGER, primaryKey: true, autoIncrement: true })
   id: number;
 
@@ -29,11 +25,8 @@ export default class Snapshot extends HiscoreValues {
   @Column({ type: DataType.INTEGER, allowNull: false, onDelete: 'CASCADE' })
   playerId: number;
 
-  @Column({ type: DataType.DATE })
-  importedAt: Date;
-
-  @CreatedAt
-  createdAt: Date;
+  @UpdatedAt
+  updatedAt: Date;
 
   /* Associations */
