+++ server/src/database/models/index.ts
@@ -1,0 +1,25 @@
+import Achievement from './Achievement';
+import Competition from './Competition';
+import Group from './Group';
+import HiscoresValues from './HiscoresValues';
+import InitialValues from './InitialValues';
+import Membership from './Membership';
+import NameChange from './NameChange';
+import Participation from './Participation';
+import Player from './Player';
+import Record from './Record';
+import Snapshot from './Snapshot';
+
+export {
+  NameChange,
+  Achievement,
+  Competition,
+  Participation,
+  HiscoresValues,
+  Player,
+  Group,
+  Membership,
+  Record,
+  Snapshot,
+  InitialValues
+};
