+++ server/tsconfig.json
@@ -11,9 +11,12 @@
     "importHelpers": true,
     "baseUrl": ".",
     "paths": {
+      "@types": ["./src/types"],
       "@controllers/*": ["./src/api/controllers/*"],
       "@services/*": ["./src/api/services/*"],
-      "@routes/*": ["./src/api/routes/*"]
+      "@routes/*": ["./src/api/routes/*"],
+      "@database": ["./src/database"],
+      "@models": ["./src/database/models"]
     },
     "lib": ["es6", "es2017", "esnext"]
   },
